Moved net payloads for ParamInterface and GlobalAnomalyStats from net header to where their respective classes are defined
