+++ include/chimbuko/net.hpp
@@ -1,7 +1,5 @@
 #pragma once
-#include "chimbuko/global_anomaly_stats.hpp"
 #include "chimbuko/util/threadPool.hpp"
-#include "chimbuko/param.hpp"
 #include "chimbuko/message.hpp"
 #include <string>
 #include <thread>
@@ -55,41 +53,6 @@ namespace chimbuko {
       check(message);
       response.set_msg(std::string("Hello!I am NET!"), false);
     };
-  };
-  class NetPayloadUpdateParams: public NetPayloadBase{
-    ParamInterface * m_param;
-  public:
-    NetPayloadUpdateParams(ParamInterface *param): m_param(param){}
-    MessageKind kind() const{ return MessageKind::PARAMETERS; }
-    MessageType type() const{ return MessageType::REQ_ADD; }
-    void action(Message &response, const Message &message) override{
-      check(message);
-      response.set_msg(m_param->update(message.buf(), true), false);
-    }	
-  };
-  class NetPayloadGetParams: public NetPayloadBase{
-    ParamInterface const* m_param;
-  public:
-    NetPayloadGetParams(ParamInterface const* param): m_param(param){}
-    MessageKind kind() const{ return MessageKind::PARAMETERS; }
-    MessageType type() const{ return MessageType::REQ_GET; }
-    void action(Message &response, const Message &message) override{
-      check(message);
-      response.set_msg(m_param->serialize(), false);
-    }	
-  };
-  class NetPayloadUpdateAnomalyStats: public NetPayloadBase{
-    GlobalAnomalyStats * m_global_anom_stats;
-  public:
-    NetPayloadUpdateAnomalyStats(GlobalAnomalyStats * global_anom_stats): m_global_anom_stats(global_anom_stats){}
-    MessageKind kind() const{ return MessageKind::ANOMALY_STATS; }
-    MessageType type() const{ return MessageType::REQ_ADD; }
-    void action(Message &response, const Message &message) override{
-      check(message);
-      if(m_global_anom_stats == nullptr) throw std::runtime_error("Cannot update global anomaly statistics as stats object has not been linked");
-      m_global_anom_stats->add_anomaly_data(message.buf());
-      response.set_msg("", false);
-    }
   };
